budget/backend/apps/api/v1/budget_api/serializers.py
from rest_framework import serializers
from budget.backend.apps.budget.models import Statement, Category


class StatementSerializer(serializers.ModelSerializer):
    class Meta:
        model = Statement
        fields = "__all__"


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = "__all__"
