budget/pyproject.toml
[tool.poetry]
name = "budget"
version = "0.1.0"
description = "simple budget app for personal use"
authors = ["VF"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django='4.2.4'
chardet='5.2.0'
charset-normalizer='3.3.2'
django-split-settings='1.2.0'
djangorestframework='3.14.0'
numpy='1.26.2'
pandas='2.1.4'
poetry='1.7.1'
poetry-core='1.8.1'
poetry-plugin-export='1.6.0'
python-magic='0.4.27'
requests='2.31.0'
requests-toolbelt='1.0.0'
virtualenv='20.25.0'
drf-spectacular='0.27.1'
django-rest-swagger='2.2.0'
argcomplete='3.2.2'
asgiref='3.7.2'
attrs='23.2.0'
build='1.0.3'
CacheControl='0.13.1'
certifi='2023.11.17'
cleo='2.1.0'
click='8.1.7'
colorama='0.4.6'
coreapi='2.3.3'
coreschema='0.0.4'
crashtest='0.4.1'
distlib='0.3.8'
dulwich='0.21.7'
et-xmlfile='1.1.0'
fastjsonschema='2.19.1'
filelock='3.13.1'
idna='3.6'
importlib-metadata='7.0.1'
inflection='0.5.1'
installer='0.7.0'
itypes='1.2.0'
Jinja2='3.1.3'
jsonschema='4.21.1'
jsonschema-specifications='2023.12.1'
keyring='24.3.0'
MarkupSafe='2.1.4'
more-itertools='10.2.0'
msgpack='1.0.7'
openapi-codec='1.3.2'
openpyxl='3.1.2'
optional-django='0.3.0'
packaging='23.2'
pexpect='4.9.0'
pipx='1.4.3'
pkginfo='1.9.6'
platformdirs='3.11.0'
psycopg2='2.9.7'
ptyprocess='0.7.0'
pyproject_hooks='1.0.0'
python-dateutil='2.8.2'
pytz='2023.3.post1'
pywin32-ctypes='0.2.2'
PyYAML='6.0.1'
rapidfuzz='3.6.1'
react='4.3.0'
referencing='0.33.0'
rpds-py='0.17.1'
shellingham='1.5.4'
simplejson='3.19.2'
six='1.16.0'
sqlparse='0.4.4'
tomli='2.0.1'
tomlkit='0.12.3'
trove-classifiers='2024.1.31'
typing_extensions='4.7.1'
tzdata='2023.3'
uritemplate='4.1.1'
urllib3='2.1.0'
userpath='1.9.1'
xlrd='2.0.1'
zipp='3.17.0'
black = "^24.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"